# api/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: "Metadata Server API"
  version: "0.4.0.0"
servers:
  - url: https://tokens.cardano.org/
    description: Production server (uses live data)
  - url: https://test.tokens.cardano.org/
    description: Test server (using live data but only read only)
...
paths:
  /metadata/{subject}:
    get:
      summary: "Query All Properties"
      tags:
        - v1 queries
      operationId: "getAllPropertiesForSubject"
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Property'
        404:
          description: '`subject` not found'
          content: {}
  /metadata/{subject}/properties/{properties}:
    get:
      summary: "Query Specific Property"
      tags:
        - v1 queries
      operationId: "getPropertyForSubject"
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
        - name: properties
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Property'
        404:
          description: '`subject` or `properties` not found'
          content: {}
  /metadata/query:
    post:
      tags:
        - v1 queries
      summary: "Batch Metadata Query"
      operationId: "getSubjects"
      requestBody:
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/BatchRequest'
        required: true
      responses:
        200:
          description: ""
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        400:
          description: Invalid `body`
          content: {}
      x-codegen-request-body-name: body
  /metadata/sync:
    post:
      tags:
        - v1 queries
      summary: "Sync Metadata Query"
      operationId: "getAll"
      requestBody:
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/BatchRequest'
        required: true
      responses:
        200:
          description: ""
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        400:
          description: Invalid `body`
          content: {}
      x-codegen-request-body-name: body      
...
components:
  schemas:
...
    NameProperty:
      required:
        - signatures
        - sequenceNumber
        - value
      type: object
      properties:
        signatures:
          type: array
          items:
            $ref: '#/components/schemas/AnnotatedSignature'
        sequenceNumber:
          type: number
          minimum: 0
        value:
          type: string
          minLength: 1
          maxLength: 50
    DescriptionProperty:
      required:
        - signatures
        - sequenceNumber
        - value
      type: object
      properties:
        signatures:
          type: array
          items:
            $ref: '#/components/schemas/AnnotatedSignature'
        sequenceNumber:
          type: number
          minimum: 0
        value:
          type: string
          maxLength: 500
    UrlProperty:
      required:
        - signatures
        - sequenceNumber
        - value
      type: object
      properties:
        signatures:
          type: array
          items:
            $ref: '#/components/schemas/AnnotatedSignature'
        sequenceNumber:
          type: number
          minimum: 0
        value:
          type: string
          maxLength: 250
          pattern: '^https://'
          example: "https://www.iohk.io"
    ToolProperty:
      type: object
      properties:
        url:
          type: string
          maxLength: 250
          pattern: '^https://'
          example: "https://www.iohk.io"
        description:
          type: string
          maxLength: 500
    TickerProperty:
      required:
        - signatures
        - sequenceNumber
        - value
      type: object
      properties:
        signatures:
          type: array
          items:
            $ref: '#/components/schemas/AnnotatedSignature'
        sequenceNumber:
          type: number
          minimum: 0
        value:
          type: string
          minLength: 2
          maxLength: 5
          example: "QUID"
    DecimalsProperty:
      required:
        - signatures
        - sequenceNumber
        - value
      type: object
      properties:
        signatures:
          type: array
          items:
            $ref: '#/components/schemas/AnnotatedSignature'
        sequenceNumber:
          type: number
          minimum: 0
        value:
          type: number
          minimum: 0
          maximum: 255
          example: 1
...
    AnnotatedSignature:
      required:
        - publicKey
        - signature
      type: object
      properties:
        signature:
          type: string
        publicKey:
          type: string
    BatchResponse:
      required:
        - subjects
      type: object
      properties:
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/Property'
    BatchRequest:
      required:
        - subjects
      type: object
      properties:
        subjects:
          type: array
          items:
            type: string
        properties:
          type: array
          items:
            type: string
...
    Property:
      type: object
      required:
        - subject
      properties:
        subject:
          type: string
          example: "b377d03a568dde663534e040fc32a57323ec00970df0e863eba3f098717569640a"
        policy:
          type: string
          example: "82008200581ce62601e8eeec975f3f124a288cd0ecb2973f5fc225629f1401a79b16"
        name:
          $ref: '#/components/schemas/NameProperty'
        description:
          $ref: '#/components/schemas/DescriptionProperty'
        url:
          $ref: '#/components/schemas/UrlProperty'
        ticker:
          $ref: '#/components/schemas/TickerProperty'
        decimals:
          $ref: '#/components/schemas/DecimalsProperty'
        logo:
          $ref: '#/components/schemas/LogoProperty'
        tool:
          $ref: '#/components/schemas/ToolProperty'